rm underscore from slideshow.
Doing this required splitting some files up. No big deal.
